get account transfer

+++ app/src/main/java/app/eospocket/android/ui/main/token/items/TokenItem.java
@@ -11,5 +11,5 @@ public class TokenItem {
 
     private String name;
 
-    private float balance;
+    private double balance;
 }
